Separate into two endpoints
