+++ langchain/scripts/create-entrypoints.js
@@ -76,7 +76,8 @@ const entrypoints = {
   "llms/googlepalm": "llms/googlepalm",
   "llms/fireworks": "llms/fireworks",
   "llms/sagemaker_endpoint": "llms/sagemaker_endpoint",
-  "llms/bedrock": "llms/bedrock",
+  "llms/bedrock": "llms/bedrock/index",
+  "llms/bedrock/web": "llms/bedrock/web",
   "llms/llama_cpp": "llms/llama_cpp",
   "llms/writer": "llms/writer",
   "llms/portkey": "llms/portkey",
@@ -185,7 +186,8 @@ const entrypoints = {
   "chat_models/openai": "chat_models/openai",
   "chat_models/portkey": "chat_models/portkey",
   "chat_models/anthropic": "chat_models/anthropic",
-  "chat_models/bedrock": "chat_models/bedrock",
+  "chat_models/bedrock": "chat_models/bedrock/index",
+  "chat_models/bedrock/web": "chat_models/bedrock/web",
   "chat_models/cloudflare_workersai": "chat_models/cloudflare_workersai",
   "chat_models/googlevertexai": "chat_models/googlevertexai/index",
   "chat_models/googlevertexai/web": "chat_models/googlevertexai/web",
@@ -352,6 +354,7 @@ const requiresOptionalDependency = [
   "llms/replicate",
   "llms/sagemaker_endpoint",
   "llms/bedrock",
+  "llms/bedrock/web",
   "llms/llama_cpp",
   "llms/writer",
   "llms/portkey",
@@ -426,6 +429,7 @@ const requiresOptionalDependency = [
   "document_transformers/html_to_text",
   "document_transformers/mozilla_readability",
   "chat_models/bedrock",
+  "chat_models/bedrock/web",
   "chat_models/googlevertexai",
   "chat_models/googlevertexai/web",
   "chat_models/googlepalm",
