+++ langchain/src/chat_models/bedrock/web.ts
@@ -9,12 +9,74 @@ import {
   BaseBedrockInput,
   BedrockLLMInputOutputAdapter,
   type CredentialType,
-} from "../util/bedrock.js";
-import { getEnvironmentVariable } from "../util/env.js";
-import { LLM, BaseLLMParams } from "./base.js";
-import { CallbackManagerForLLMRun } from "../callbacks/manager.js";
-import { GenerationChunk } from "../schema/index.js";
-import { SerializedFields } from "../load/map_keys.js";
+} from "../../util/bedrock.js";
+import { getEnvironmentVariable } from "../../util/env.js";
+import { SimpleChatModel, BaseChatModelParams } from "../base.js";
+import { CallbackManagerForLLMRun } from "../../callbacks/manager.js";
+import {
+  AIMessageChunk,
+  BaseMessage,
+  AIMessage,
+  ChatGenerationChunk,
+  ChatMessage,
+} from "../../schema/index.js";
+import { SerializedFields } from "../../load/map_keys.js";
+
+function convertOneMessageToText(
+  message: BaseMessage,
+  humanPrompt: string,
+  aiPrompt: string
+): string {
+  if (message._getType() === "human") {
+    return `${humanPrompt} ${message.content}`;
+  } else if (message._getType() === "ai") {
+    return `${aiPrompt} ${message.content}`;
+  } else if (message._getType() === "system") {
+    return `${humanPrompt} <admin>${message.content}</admin>`;
+  } else if (ChatMessage.isInstance(message)) {
+    return `\n\n${
+      message.role[0].toUpperCase() + message.role.slice(1)
+    }: {message.content}`;
+  }
+  throw new Error(`Unknown role: ${message._getType()}`);
+}
+
+export function convertMessagesToPromptAnthropic(
+  messages: BaseMessage[],
+  humanPrompt = "\n\nHuman:",
+  aiPrompt = "\n\nAssistant:"
+): string {
+  const messagesCopy = [...messages];
+
+  if (
+    messagesCopy.length === 0 ||
+    messagesCopy[messagesCopy.length - 1]._getType() !== "ai"
+  ) {
+    messagesCopy.push(new AIMessage({ content: "" }));
+  }
+
+  return messagesCopy
+    .map((message) => convertOneMessageToText(message, humanPrompt, aiPrompt))
+    .join("");
+}
+
+/**
+ * Function that converts an array of messages into a single string prompt
+ * that can be used as input for a chat model. It delegates the conversion
+ * logic to the appropriate provider-specific function.
+ * @param messages Array of messages to be converted.
+ * @param options Options to be used during the conversion.
+ * @returns A string prompt that can be used as input for a chat model.
+ */
+export function convertMessagesToPrompt(
+  messages: BaseMessage[],
+  provider: string
+): string {
+  if (provider === "anthropic") {
+    return convertMessagesToPromptAnthropic(messages);
+  }
+  throw new Error(`Provider ${provider} does not support chat.`);
+}
 
 /**
  * A type of Large Language Model (LLM) that interacts with the Bedrock
@@ -25,7 +87,7 @@ import { SerializedFields } from "../load/map_keys.js";
  * configured with various parameters such as the model to use, the AWS
  * region, and the maximum number of tokens to generate.
  */
-export class Bedrock extends LLM implements BaseBedrockInput {
+export class BedrockChat extends SimpleChatModel implements BaseBedrockInput {
   model = "amazon.titan-tg1-large";
 
   region: string;
@@ -74,10 +136,10 @@ export class Bedrock extends LLM implements BaseBedrockInput {
   }
 
   static lc_name() {
-    return "Bedrock";
+    return "BedrockChat";
   }
 
-  constructor(fields?: Partial<BaseBedrockInput> & BaseLLMParams) {
+  constructor(fields?: Partial<BaseBedrockInput> & BaseChatModelParams) {
     super(fields ?? {});
 
     this.model = fields?.model ?? this.model;
@@ -124,7 +186,7 @@ export class Bedrock extends LLM implements BaseBedrockInput {
       response = model.call("Tell me a joke.")
   */
   async _call(
-    prompt: string,
+    messages: BaseMessage[],
     options: this["ParsedCallOptions"],
     runManager?: CallbackManagerForLLMRun
   ): Promise<string> {
@@ -133,8 +195,8 @@ export class Bedrock extends LLM implements BaseBedrockInput {
       this.endpointHost ?? `${service}.${this.region}.amazonaws.com`;
     const provider = this.model.split(".")[0];
     if (this.streaming) {
-      const stream = this._streamResponseChunks(prompt, options, runManager);
-      let finalResult: GenerationChunk | undefined;
+      const stream = this._streamResponseChunks(messages, options, runManager);
+      let finalResult: ChatGenerationChunk | undefined;
       for await (const chunk of stream) {
         if (finalResult === undefined) {
           finalResult = chunk;
@@ -142,9 +204,10 @@ export class Bedrock extends LLM implements BaseBedrockInput {
           finalResult = finalResult.concat(chunk);
         }
       }
-      return finalResult?.text ?? "";
+      return finalResult?.message.content ?? "";
     }
-    const response = await this._signedFetch(prompt, options, {
+
+    const response = await this._signedFetch(messages, options, {
       bedrockMethod: "invoke",
       endpointHost,
       provider,
@@ -160,7 +223,7 @@ export class Bedrock extends LLM implements BaseBedrockInput {
   }
 
   async _signedFetch(
-    prompt: string,
+    messages: BaseMessage[],
     options: this["ParsedCallOptions"],
     fields: {
       bedrockMethod: "invoke" | "invoke-with-response-stream";
@@ -171,7 +234,7 @@ export class Bedrock extends LLM implements BaseBedrockInput {
     const { bedrockMethod, endpointHost, provider } = fields;
     const inputBody = BedrockLLMInputOutputAdapter.prepareInput(
       provider,
-      prompt,
+      convertMessagesToPromptAnthropic(messages),
       this.maxTokens,
       this.temperature,
       options.stop ?? this.stopSequences,
@@ -220,34 +283,23 @@ export class Bedrock extends LLM implements BaseBedrockInput {
     return response;
   }
 
-  invocationParams(options?: this["ParsedCallOptions"]) {
-    return {
-      model: this.model,
-      region: this.region,
-      temperature: this.temperature,
-      maxTokens: this.maxTokens,
-      stop: options?.stop ?? this.stopSequences,
-      modelKwargs: this.modelKwargs,
-    };
-  }
-
   async *_streamResponseChunks(
-    prompt: string,
+    messages: BaseMessage[],
     options: this["ParsedCallOptions"],
     runManager?: CallbackManagerForLLMRun
-  ): AsyncGenerator<GenerationChunk> {
+  ): AsyncGenerator<ChatGenerationChunk> {
     const provider = this.model.split(".")[0];
+    const service = "bedrock-runtime";
+
+    const endpointHost =
+      this.endpointHost ?? `${service}.${this.region}.amazonaws.com`;
+
     const bedrockMethod =
       provider === "anthropic" || provider === "cohere"
         ? "invoke-with-response-stream"
         : "invoke";
 
-    const service = "bedrock-runtime";
-    const endpointHost =
-      this.endpointHost ?? `${service}.${this.region}.amazonaws.com`;
-
-    // Send request to AWS using the low-level fetch API
-    const response = await this._signedFetch(prompt, options, {
+    const response = await this._signedFetch(messages, options, {
       bedrockMethod,
       endpointHost,
       provider,
@@ -287,9 +339,9 @@ export class Bedrock extends LLM implements BaseBedrockInput {
             provider,
             chunkResult
           );
-          yield new GenerationChunk({
+          yield new ChatGenerationChunk({
             text,
-            generationInfo: {},
+            message: new AIMessageChunk({ content: text }),
           });
           // eslint-disable-next-line no-void
           void runManager?.handleLLMNewToken(text);
@@ -298,9 +350,9 @@ export class Bedrock extends LLM implements BaseBedrockInput {
     } else {
       const json = await response.json();
       const text = BedrockLLMInputOutputAdapter.prepareOutput(provider, json);
-      yield new GenerationChunk({
+      yield new ChatGenerationChunk({
         text,
-        generationInfo: {},
+        message: new AIMessageChunk({ content: text }),
       });
       // eslint-disable-next-line no-void
       void runManager?.handleLLMNewToken(text);
@@ -318,5 +370,9 @@ export class Bedrock extends LLM implements BaseBedrockInput {
         }
       },
     };
+  }
+
+  _combineLLMOutput() {
+    return {};
   }
 }
