+++ langchain/src/llms/bedrock/index.ts
@@ -1,0 +1,17 @@
+import { defaultProvider } from "@aws-sdk/credential-provider-node";
+import { BaseBedrockInput } from "../../util/bedrock.js";
+import { BaseLLMParams } from "../base.js";
+import { Bedrock as BaseBedrock } from "./web.js";
+
+export class Bedrock extends BaseBedrock {
+  static lc_name() {
+    return "Bedrock";
+  }
+
+  constructor(fields?: Partial<BaseBedrockInput> & BaseLLMParams) {
+    super({
+      ...fields,
+      credentials: fields?.credentials ?? defaultProvider(),
+    });
+  }
+}
