+++ langchain/tsconfig.json
@@ -93,7 +93,8 @@
       "src/llms/googlepalm.ts",
       "src/llms/fireworks.ts",
       "src/llms/sagemaker_endpoint.ts",
-      "src/llms/bedrock.ts",
+      "src/llms/bedrock/index.ts",
+      "src/llms/bedrock/web.ts",
       "src/llms/llama_cpp.ts",
       "src/llms/writer.ts",
       "src/llms/portkey.ts",
@@ -187,7 +188,8 @@
       "src/chat_models/openai.ts",
       "src/chat_models/portkey.ts",
       "src/chat_models/anthropic.ts",
-      "src/chat_models/bedrock.ts",
+      "src/chat_models/bedrock/index.ts",
+      "src/chat_models/bedrock/web.ts",
       "src/chat_models/cloudflare_workersai.ts",
       "src/chat_models/googlevertexai/index.ts",
       "src/chat_models/googlevertexai/web.ts",
